requirements.txt
twisted==22.10.0
PyYAML==6.0
redis==4.4.2
flask==2.2.2
flask-cors==3.0.10
Flask-APScheduler==1.12.4
flask-redis==0.4.0
requests==2.28.2
pyjwt==2.6.0
service_identity==21.1.0